Fix compiler warnings

+++ src/comb.cpp
@@ -150,7 +150,7 @@ void print_timer(CommInfo& comm_info, Timer& tm, const char* prefix = "") {
   double* maxs = new double[res.size()];
   long  * nums = new long  [res.size()];
 
-  for (int i = 0; i < res.size(); ++i) {
+  for (int i = 0; i < (int)res.size(); ++i) {
     sums[i] = res[i].sum;
     mins[i] = res[i].min;
     maxs[i] = res[i].max;
@@ -175,7 +175,7 @@ void print_timer(CommInfo& comm_info, Timer& tm, const char* prefix = "") {
 
   if (comm_info.rank == 0) {
 
-    for (int i = 0; i < res.size(); ++i) {
+    for (int i = 0; i < (int)res.size(); ++i) {
       int padding = max_name_len - res[i].name.size();
       comm_info.print(FileGroup::summary, "%s%s:%*s num %ld sum %.9f s min %.9f s max %.9f s\n",
                              prefix, res[i].name.c_str(), padding, "", final_nums[i], final_sums[i], final_mins[i], final_maxs[i]);
@@ -187,7 +187,7 @@ void print_timer(CommInfo& comm_info, Timer& tm, const char* prefix = "") {
     delete[] final_nums;
   }
 
-  for (int i = 0; i < res.size(); ++i) {
+  for (int i = 0; i < (int)res.size(); ++i) {
     int padding = max_name_len - res[i].name.size();
     comm_info.print(FileGroup::proc, "%s%s:%*s num %ld sum %.9f s min %.9f s max %.9f s\n",
                         prefix, res[i].name.c_str(), padding, "", nums[i], sums[i], mins[i], maxs[i]);
